DEV: Code cleanup
* move the component code (and the template) into the plugin connector
* use `shouldRender`
* instead of getter with side-effects, use `did-insert`
* `this.` in the template
