+++ javascripts/discourse/connectors/header-list-container-bottom/category-featured-images.hbs
@@ -1,0 +1,34 @@
+<DSection
+  {{did-insert this.showFooter}}
+  @pageClass={{if
+    this.hideTopicList
+    "category-featured-images--hide-topic-list"
+  }}
+/>
+
+{{#each this.filteredSetting as |fs|}}
+  <div class="category-featured-image-section custom-filter_{{fs.slug}}">
+    {{#if fs.section_title}}
+      <h2 class="category-featured-image-section__title">
+        {{fs.section_title}}
+      </h2>
+    {{/if}}
+
+    <div class="category-featured-image-section__images">
+      {{#each fs.images as |image|}}
+        <a href={{image.link}}>
+          <img
+            class="category-featured-image-section__image"
+            src={{image.image_url}}
+          />
+
+          {{#if image.link_text}}
+            <h3 class="category-featured-image-section__image-title">
+              {{image.link_text}}
+            </h3>
+          {{/if}}
+        </a>
+      {{/each}}
+    </div>
+  </div>
+{{/each}}
