--- javascripts/discourse/connectors/header-list-container-bottom/category-image-connector.hbs
@@ -1,1 +1,0 @@
-<CategoryFeaturedImages @category={{this.category}} />
